chore: remove unnecessary component
